chore: move socket-ct to server
