+++ packages/server-ct/src/server-ct.ts
@@ -2,7 +2,7 @@ import Bluebird from 'bluebird'
 import httpsProxy from '@packages/https-proxy'
 import { OpenServerOptions, ServerBase } from '@packages/server/lib/server-base'
 import appData from '@packages/server/lib/util/app_data'
-import type { SocketCt } from './socket-ct'
+import type { SocketCt } from '@packages/server/lib/socket-ct'
 import type { Cfg } from '@packages/server/lib/project-base'
 
 type WarningErr = Record<string, any>
