+++ packages/server/lib/project-base.ts
@@ -6,8 +6,9 @@ import path from 'path'
 
 import browsers from './browsers'
 import pkg from '@packages/root'
-import { ServerCt, SocketCt } from '@packages/server-ct'
+import { ServerCt } from '@packages/server-ct'
 import { SocketE2E } from './socket-e2e'
+import { SocketCt } from './socket-ct'
 import api from './api'
 import { Automation } from './automation'
 import * as config from './config'
