+++ packages/server/lib/server-base.ts
@@ -14,7 +14,7 @@ import type httpsProxy from '@packages/https-proxy'
 import { netStubbingState, NetStubbingState } from '@packages/net-stubbing'
 import { agent, clientCertificates, cors, httpUtils, uri } from '@packages/network'
 import { NetworkProxy, BrowserPreRequest } from '@packages/proxy'
-import type { SocketCt } from '@packages/server-ct'
+import type { SocketCt } from './socket-ct'
 import errors from './errors'
 import logger from './logger'
 import Request from './request'
